chore(): correct interfaces for hash and browser history.
