+++ src/global/interfaces.ts
@@ -5,3 +5,26 @@ export interface ActiveRouter {
 }
 
 export type Listener = () => void;
+
+export interface LocationSegments {
+  pathname?: string;
+  search?: string;
+  hash?: string;
+  state?: any;
+  key?: string;
+}
+
+export interface RouterHistory {
+  length: number;
+  action: string;
+  location: LocationSegments;
+  createHref: (location: LocationSegments) => string;
+  push: (path: string | LocationSegments, state: any) => void;
+  replace: (path: string | LocationSegments, state: any) => void;
+  go: (n: number) => void;
+  goBack: () => void;
+  goForward: () => void;
+  block: (prompt?: string) => () => void;
+  listen: (listener: Function) => () => void;
+}
+
